@@ -10,11 +10,8 @@ license = {file = "LICENSE"}
 classifiers = ["License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)"]
 dynamic = ["version", "description"]
 keywords = ["museum", "museumPlus", "api client"]
-# datetime should be included in Python
-# flit?
-#requires-python = 3.9.1 
+requires-python = 3.11 # union types with | 
 dependencies = [
-	#"dateutil", 
 	"flit",
 	"lxml >=4.8.0", 
 	"openpyxl",
